Fix issue #5 Provide richer API for modules
Provide a mechanism to download documents using Weboob modules
abilities.

If a user wants to download any document, it will be fetched in a
temporary folder. The file URI will be passed back in the output JSON.

If the users then have access to the server filesystem, they can
directly get the requested document and remove temporary files when they
are no longer needed.

If the users do not have such access, they can use the `/retrieve`
endpoint exposed by the webserver to retrieve the content of this
file. Note that this endpoint is not designed for a production
environment and might expose other sensitive content from your temporary
directory. Note also that this endpoint will note delete the temporary
file.

Closes #5.